refactor: split app history and goal helpers
